Initial attempt at local+remote integration
This commit is a mostly-working implementation of an integrated
remote+local setup. Details:
- Local setup relies on MDNS/Zeroconf discovery; there is no support for
  statically specifying hosts.
- Local and Remote devices are currently set up as the same
  homeassistant device, which is annoying for various reasons.

The intended flow is that a user either:
- Has never configured remote support at all, and relies solely on
  zeroconf discovery and the config flow kicked off from discovery
- Has configured remote support before, but has disabled remote devices
  they wish to poll locally *before* entering the config flow kicked off
  from discovery
- Has removed the instance of the integration configured with remote
  support entirely

It's a little messy if you set up both local *and* remote
simultaneously for the same Awair device. It *works*! But you can't ever
really disable remote and/or local support once you get into that state
- you're stuck with both until you remove the integration.

I think this is just a starting point, but I wanted to push something up
for consideration before I went too much further. What I'd like to do
next is:
- Track remote + local device UUIDs in config_entries somehow
- Set up an Options flow for the integration so that you can toggle
  remote/local polling for devices that can do both
- Automatically upgrade config entries to track preferences for
  remote/local polling when the integrations are set up.

The major barriers to improvements are:
- The discovery flow is not ever supposed to modify configuration
  without consent, which makes some things difficult (or really, just
  means we have to set up more UI flow configs and I dislike doing
  programmatic UI configs like this!)
- We're not really guaranteed that setup/discovery will run in any
  specific order, but we can probably not worry about that too much.
- There doesn't seem to be much of a pattern within HA so far of
  updating config entries during platform setup. I'm worried we'd be
  doing something that's not implicitly supported.

If we can overcome the usability problems, then I think this has a
really nice future. Otherwise I think we should actually revert and go
the route of two separate integrations (but as expressed elsewhere, I
think that has its own flaws).

If you pull this into your HA installation, you'll need to modify the
`generated/zeroconf.py` file like so:

```diff
@@ -75,6 +75,10 @@ ZEROCONF = {
         }
     ],
     "_http._tcp.local.": [
+        {
+            "domain": "awair",
+            "name": "awair*"
+        },
         {
             "domain": "shelly",
             "name": "shelly*"
```

..otherwise HA will never discover the Awair.